SER-656 Post feature refactoring

+++ modules/front-end/app/models/frontend/storefront/ChoosePhotoTileProduct.scala
@@ -16,6 +16,6 @@ case class ChoosePhotoTileProduct(
   price: Money,
   imageUrl: String,
   targetUrl: String,
-  quantityRemaining: Int,
+  quantityRemaining: BigInt,
   orientation: ProductOrientation
 )
